+++ src/main/kotlin/io/runnershigh/backend/training/entity/TrainingSchedules.kt
@@ -2,7 +2,7 @@ package io.runnershigh.backend.training.entity
 
 import io.runnershigh.backend.shared.entity.BaseEntity
 import io.runnershigh.backend.training.entity.enum.TrainingStatus
-import io.runnershigh.backend.user.infrastructure.entity.UserEntity
+import io.runnershigh.backend.user.entity.UserEntity
 import jakarta.persistence.*
 import org.hibernate.annotations.Comment
 import java.time.LocalDate
