+++ src/main/kotlin/io/runnershigh/backend/training/mapper/TrainingScheduleMapper.kt
@@ -3,7 +3,7 @@ package io.runnershigh.backend.training.mapper
 import io.runnershigh.backend.training.entity.TrainingSchedules
 import io.runnershigh.backend.training.dto.request.SaveTrainingSchedule
 import io.runnershigh.backend.training.dto.response.ReadTrainingSchedule
-import io.runnershigh.backend.user.infrastructure.entity.UserEntity
+import io.runnershigh.backend.user.entity.UserEntity
 
 // request
 fun SaveTrainingSchedule.toEntity(user: UserEntity) = TrainingSchedules(
