+++ src/main/kotlin/io/runnershigh/backend/user/entity/enum/UserStatus.kt
@@ -1,4 +1,4 @@
-package io.runnershigh.backend.user.domain.enum
+package io.runnershigh.backend.user.entity.enum
 
 enum class UserStatus(var value: String) {
     ACTIVE("정상"),
